fix: log worker deployment config with json info
Using fmt.Println results in the config being spread out across many
log lines as the config is a multi-line string and \n is interpreted as
new log lines. Additionally, using fmt.Println means that we're missing
out on any added labels like processInstanceKey or clusterId.

We can change this by using LogInfo or LogVerbose. I've chosen LogInfo
here as we can log the config together with a stronger statement that
we're deploying the worker now.