feat: added new test case for dynamic scaling
Similar to broker-scaling.json, but the number of partitions is scaled
to 4 (on top of broker scaling).

Since new partitions do not participate in message correlation, there is
no test in verifying that messages are published to the new partition.
However, a new process instance is created on the new partition,
receiving a message from another partition.

Brokers are scaled back to 3 when done, but partitions cannot be scaled
back.